1. Fix an error when fetch string type field may cause malform packet error. (#7262)
This is beacuse of an const MAX_PHYSICAL_PACKET_LENGTH  in fe should be 2^24 -1,
   but it is set as 2^24 -2 by mistake.
2. Fix bitmap_to_string may failed when the result is large than 2G

+++ be/src/udf/udf.h
@@ -638,14 +638,14 @@ struct StringVal : public AnyVal {
     /// Will create a new StringVal with the given dimension and copy the data from the
     /// parameters. In case of an error will return a nullptr string and set an error on the
     /// function context.
-    static StringVal copy_from(FunctionContext* ctx, const uint8_t* buf, size_t len);
+    static StringVal copy_from(FunctionContext* ctx, const uint8_t* buf, int64_t len);
 
     /// Append the passed buffer to this StringVal. Reallocate memory to fit the buffer. If
     /// the memory allocation becomes too large, will set an error on FunctionContext and
     /// return a nullptr string.
-    void append(FunctionContext* ctx, const uint8_t* buf, size_t len);
-    void append(FunctionContext* ctx, const uint8_t* buf, size_t len, const uint8_t* buf2,
-                size_t buf2_len);
+    void append(FunctionContext* ctx, const uint8_t* buf, int64_t len);
+    void append(FunctionContext* ctx, const uint8_t* buf, int64_t len, const uint8_t* buf2,
+                int64_t buf2_len);
 };
 
 struct DecimalV2Val : public AnyVal {
